requirements.txt
# Minimal requirements - install one by one if needed
fastapi
uvicorn
python-multipart
openai
python-dotenv
requests
numpy
PyMuPDF
faiss-cpu